chore: move rendering errors in the top of canvas
