+++ packages/server/next/app/components/ui/chat/chat-message-content.tsx
@@ -8,10 +8,8 @@ import { ToolAnnotations } from "./tools/chat-tools";
 
 export function ChatMessageContent({
   componentDefs,
-  appendError,
 }: {
   componentDefs: ComponentDef[];
-  appendError: (error: string) => void;
 }) {
   return (
     <ChatMessage.Content>
@@ -19,7 +17,7 @@ export function ChatMessageContent({
       <ChatMessage.Content.AgentEvent />
       <ToolAnnotations />
       <ChatMessage.Content.Image />
-      <DynamicEvents componentDefs={componentDefs} appendError={appendError} />
+      <DynamicEvents componentDefs={componentDefs} />
       <ArtifactCard />
       <ChatMessage.Content.Markdown />
       <ChatMessage.Content.DocumentFile />
